pyproject.toml
[project]
name = "my-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "copilotkit>=0.1.79",
    "langchain-core>=1.2.19",
    "langchain-openai>=1.1.11",
    "langgraph>=1.0.10",
    "python-dotenv>=1.0.1",
    "tavily-python>=0.7.23",
]

[dependency-groups]
dev = [
    "mypy>=1.14.0",
    "pre-commit>=4.0.0",
    "ruff>=0.9.0",
]

[tool.ruff]
target-version = "py312"
line-length = 100
src = ["."]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
    "SIM", # flake8-simplify
]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.12"
strict = false
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
check_untyped_defs = true
no_implicit_optional = true
show_error_codes = true

[[tool.mypy.overrides]]
module = [
    "copilotkit.*",
    "langchain_core.*",
    "langchain_openai.*",
    "langgraph.*",
    "tavily.*",
]
ignore_missing_imports = true
